+++ src/entities/weather/Hail.ts
