+++ tsconfig.json
@@ -18,7 +18,6 @@
     // ]
   },
   "include": [
-    "src",
-    "types.d.ts"
+    "src"
   ]
 }
